requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
transformers>=4.30.0
torch>=2.0.0
Pillow>=10.0.0
python-dotenv>=0.19.0
gtts>=2.3.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=0.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0 